bitstream.c: normalize some size types
size_t -> uint32_t, quiets -Wshorten-64-to-32 warnings

BUG=aomedia:445

+++ aom_dsp/bitreader_buffer.c
@@ -16,8 +16,8 @@ size_t aom_rb_bytes_read(struct aom_read_bit_buffer *rb) {
 }
 
 int aom_rb_read_bit(struct aom_read_bit_buffer *rb) {
-  const size_t off = rb->bit_offset;
-  const size_t p = off >> 3;
+  const uint32_t off = rb->bit_offset;
+  const uint32_t p = off >> 3;
   const int q = 7 - (int)(off & 0x7);
   if (rb->bit_buffer + p < rb->bit_buffer_end) {
     const int bit = (rb->bit_buffer[p] >> q) & 1;
